Use `Text.value()` instead of merging
